feat: Добавить Section
